requirements.txt
matplotlib
pillow
opencv-python